Refactor: Migrate state management to Zustand
This commit replaces React Context with Zustand for game and auth state management. It centralizes game mechanics, improves performance, and enhances developer experience. Backward compatibility is maintained through context wrappers.
